render-requirements.txt
fFlask==2.3.2
gunicorn==20.1.0
requests==2.31.0
python-dotenv==1.0.0
whitenoise==6.5.0
markupsafe==2.1.3
werkzeug==2.3.6
itsdangerous==2.1.2
click==8.1.3
blinker==1.6.20
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
twilio==8.10.0